Change text filter to have special behavior
The text filter will now cascade the rank of a included entity if the
depth is 1. The idea is that we want to include nested support paths if
the parent path has the matching entity. If the parent path does not
have the matching entity, then we want to look at the support paths like
normal.